% Generated by roxygen2 (4.0.0): do not edit by hand
\name{how_big_is_my_package}
\alias{how_big_is_my_package}
\title{Counts the total number of lines of R code in a package.}
\usage{
how_big_is_my_package(package_dir)
}
\arguments{
  \item{package_dir}{Package directory. Must be source
  code, not a binary.}
}
\description{
Counts the total number of lines of R code in a package.
}
\examples{
\dontrun{

how_big_is_my_package("~/Dropbox/R_Working_Directory/swirl")
}
}

